Add --homedir and NEWSBEUTERHOME env variable.
The only way to change home directories until now were the -u -c and -C
arguments, and even with these you weren't able to change the history files.
This commit implements the --homedir argument which switches all files to the
directory provided. Additionally it adds support for the NEWSBEUTERHOME
environment variable that behaves exactly as --homedir.

With this commit newsbeuter will look for its files in this order:

-u -c and -C arguments > --homedir > NEWSBEUTERHOME variable > XDG directories > $HOME/.newsbeuter