intercept browser's failure (issue #448)
When opening a link the browser may return a value unequal to zero.
This gets recognized and then printed to stderr.

If this happens during the "open-in-browser-and-mark-read"-operation neither
the article in question is marked as read nor it is scrolled down.

Unfortunately I had to change the method declaration of open_in_browser from void
to bool in order to transmit one bit of information to the outside of that
particular function. Although this should not be an issue as this is
syntactically allowed similar a line of code only persisting of a literal.